refactor: Extract scaling commands to src/commands/scaling.py (Phase 2)
Following Phase 1 database command modularization pattern (PR #545).

Changes:
- NEW: src/commands/scaling.py (825 lines)
  - scale-up group with template and scenario commands
  - scale-down group with algorithm and pattern commands
  - Individual commands: scale-clean, scale-validate, scale-stats
- NEW: tests/commands/test_scaling.py (472 lines)
  - 21 tests covering all scaling commands
  - Help text verification
  - Successful operation tests with mocked handlers
  - Backward compatibility tests
- MODIFIED: scripts/cli.py (-732 lines)
  - Import scaling commands from src/commands/scaling
  - Register command groups and individual commands
  - Remove inline implementations

Issue #427: Scaling Operations
Issue #482: CLI Modularization (Phase 2)

Line Reduction: 732 lines removed from cli.py (3134 → 2402)
Test Coverage: 21 tests passing, all help text and basic operations verified
Backward Compatibility: All command aliases and imports preserved
