+++ scripts/cli.py
@@ -28,6 +28,13 @@ from src.commands.database import (
     restore as restore_cmd,
     wipe as wipe_cmd,
 )
+from src.commands.scaling import (
+    scale_clean as scale_clean_cmd,
+    scale_down as scale_down_cmd,
+    scale_stats as scale_stats_cmd,
+    scale_up as scale_up_cmd,
+    scale_validate as scale_validate_cmd,
+)
 
 # Import modular commands for CLI registration (Issue #482)
 from src.commands.config import config as config_cmd
@@ -36,6 +43,7 @@ from src.commands.deploy import deploy_command
 from src.commands.doctor import check_permissions as check_permissions_cmd
 from src.commands.doctor import doctor as doctor_cmd
 from src.commands.export_abstraction import export_abstraction_command
+from src.commands.layer_cmd import layer as layer_group
 from src.commands.list_deployments import list_deployments
 from src.commands.spa import spa_start as spa_start_command
 from src.commands.spa import spa_stop as spa_stop_command
@@ -1082,6 +1090,9 @@ cli.add_command(check_permissions_cmd, "check-permissions")
 cli.add_command(config_cmd, "config")
 cli.add_command(container_cmd, "container")
 
+# Register layer command group (Issue #482: CLI Modularization - Phase 3)
+cli.add_command(layer_group)
+
 
 @cli.command()
 @click.option(
@@ -1569,1349 +1580,18 @@ async def cost_report(
 
 
 # ============================================================================
-# Scale Operations Commands (Issue #427)
+# Scale Operations Commands (Issue #427, Issue #482 - Phase 2)
 # ============================================================================
 
-
-@cli.group(name="scale-up")
-def scale_up() -> None:
-    """Scale up operations - add synthetic nodes to the graph for testing."""
-    pass
-
-
-@scale_up.command(name="template")
-@click.option(
-    "--tenant-id",
-    required=False,
-    help="Azure tenant ID (defaults to AZURE_TENANT_ID from .env)",
-)
-@click.option(
-    "--template-file",
-    required=True,
-    type=click.Path(exists=True, dir_okay=False, readable=True),
-    help="Path to YAML template file defining resource structure",
-)
-@click.option(
-    "--scale-factor",
-    type=float,
-    default=2.0,
-    help="Multiplication factor for resources (default: 2.0)",
-)
-@click.option(
-    "--batch-size",
-    type=int,
-    default=500,
-    help="Number of resources to process per batch (default: 500)",
-)
-@click.option(
-    "--dry-run",
-    is_flag=True,
-    help="Preview changes without executing",
-)
-@click.option(
-    "--no-validate",
-    is_flag=True,
-    help="Skip pre/post validation checks",
-)
-@click.option(
-    "--config",
-    type=click.Path(exists=True, dir_okay=False, readable=True),
-    help="Path to scale configuration file",
-)
-@click.option(
-    "--output-format",
-    type=click.Choice(["table", "json", "markdown"], case_sensitive=False),
-    default="table",
-    help="Output format for results (default: table)",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def scale_up_template(
-    ctx: click.Context,
-    tenant_id: Optional[str],
-    template_file: str,
-    scale_factor: float,
-    batch_size: int,
-    dry_run: bool,
-    no_validate: bool,
-    config: Optional[str],
-    output_format: str,
-    no_container: bool,
-) -> None:
-    """
-    Scale up using template-based generation.
-
-    Reads a YAML template defining resource types, counts, and relationships,
-    then generates synthetic nodes based on the template multiplied by the
-    scale factor.
-
-    Example:
-
-        \b
-        # Scale up using a template with 2x multiplier
-        atg scale-up template --template-file templates/small_env.yaml --scale-factor 2.0
-
-        \b
-        # Preview changes without executing
-        atg scale-up template --template-file templates/large_env.yaml --dry-run
-
-        \b
-        # Use custom configuration
-        atg scale-up template --template-file templates/test.yaml --config scale-config.yaml
-    """
-    from src.cli_commands_scale import scale_up_template_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    if not no_container:
-        ensure_neo4j_running(debug)
-
-    await scale_up_template_command_handler(
-        tenant_id=tenant_id,
-        template_file=template_file,
-        scale_factor=scale_factor,
-        batch_size=batch_size,
-        dry_run=dry_run,
-        validate=not no_validate,
-        config_path=config,
-        output_format=output_format,
-        debug=debug,
-        no_container=no_container,
-    )
-
-
-@scale_up.command(name="scenario")
-@click.option(
-    "--tenant-id",
-    required=False,
-    help="Azure tenant ID (defaults to AZURE_TENANT_ID from .env)",
-)
-@click.option(
-    "--scenario",
-    required=True,
-    type=click.Choice(
-        ["hub-spoke", "multi-region", "dev-test-prod"], case_sensitive=False
-    ),
-    help="Built-in scenario template to use",
-)
-@click.option(
-    "--scale-factor",
-    type=float,
-    default=1.0,
-    help="Multiplication factor for scenario resources (default: 1.0)",
-)
-@click.option(
-    "--regions",
-    type=str,
-    help="Comma-separated list of Azure regions (for multi-region scenario)",
-)
-@click.option(
-    "--spoke-count",
-    type=int,
-    default=3,
-    help="Number of spoke VNets (for hub-spoke scenario, default: 3)",
-)
-@click.option(
-    "--dry-run",
-    is_flag=True,
-    help="Preview changes without executing",
-)
-@click.option(
-    "--no-validate",
-    is_flag=True,
-    help="Skip pre/post validation checks",
-)
-@click.option(
-    "--config",
-    type=click.Path(exists=True, dir_okay=False, readable=True),
-    help="Path to scale configuration file",
-)
-@click.option(
-    "--output-format",
-    type=click.Choice(["table", "json", "markdown"], case_sensitive=False),
-    default="table",
-    help="Output format for results (default: table)",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def scale_up_scenario(
-    ctx: click.Context,
-    tenant_id: Optional[str],
-    scenario: str,
-    scale_factor: float,
-    regions: Optional[str],
-    spoke_count: int,
-    dry_run: bool,
-    no_validate: bool,
-    config: Optional[str],
-    output_format: str,
-    no_container: bool,
-) -> None:
-    """
-    Scale up using scenario-based generation.
-
-    Uses built-in scenario templates that represent common Azure deployment
-    patterns:
-
-    \b
-    - hub-spoke: Hub-and-spoke network topology with peered VNets
-    - multi-region: Resources distributed across multiple Azure regions
-    - dev-test-prod: Three-tier environment structure
-
-    Example:
-
-        \b
-        # Generate a hub-spoke topology with 5 spokes
-        atg scale-up scenario --scenario hub-spoke --spoke-count 5
-
-        \b
-        # Multi-region deployment across 3 regions
-        atg scale-up scenario --scenario multi-region --regions eastus,westus,northeurope
-
-        \b
-        # Dev/Test/Prod environment with 2x scale factor
-        atg scale-up scenario --scenario dev-test-prod --scale-factor 2.0
-    """
-    from src.cli_commands_scale import scale_up_scenario_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    if not no_container:
-        ensure_neo4j_running(debug)
-
-    # Parse regions if provided
-    region_list = regions.split(",") if regions else None
-
-    await scale_up_scenario_command_handler(
-        tenant_id=tenant_id,
-        scenario=scenario,
-        scale_factor=scale_factor,
-        regions=region_list,
-        spoke_count=spoke_count,
-        dry_run=dry_run,
-        validate=not no_validate,
-        config_path=config,
-        output_format=output_format,
-        debug=debug,
-        no_container=no_container,
-    )
-
-
-@cli.group(name="scale-down")
-def scale_down() -> None:
-    """Scale down operations - sample/reduce the graph for testing."""
-    pass
-
-
-@scale_down.command(name="algorithm")
-@click.option(
-    "--tenant-id",
-    required=False,
-    help="Azure tenant ID (defaults to AZURE_TENANT_ID from .env)",
-)
-@click.option(
-    "--algorithm",
-    required=True,
-    type=click.Choice(
-        ["forest-fire", "mhrw", "random-walk", "random-node"], case_sensitive=False
-    ),
-    help="Sampling algorithm to use",
-)
-@click.option(
-    "--target-size",
-    type=float,
-    default=0.1,
-    help="Target size as fraction of original (0.0-1.0, default: 0.1)",
-)
-@click.option(
-    "--target-count",
-    type=int,
-    help="Target absolute node count (overrides target-size)",
-)
-@click.option(
-    "--burn-in",
-    type=int,
-    default=50,
-    help="Burn-in steps for forest-fire algorithm (default: 50)",
-)
-@click.option(
-    "--burning-prob",
-    type=float,
-    default=0.4,
-    help="Burning probability for forest-fire (default: 0.4)",
-)
-@click.option(
-    "--walk-length",
-    type=int,
-    default=1000,
-    help="Walk length for MHRW algorithm (default: 1000)",
-)
-@click.option(
-    "--alpha",
-    type=float,
-    default=1.0,
-    help="Bias parameter for MHRW (default: 1.0 = unbiased)",
-)
-@click.option(
-    "--output-mode",
-    type=click.Choice(["delete", "export", "new-tenant"], case_sensitive=False),
-    default="delete",
-    help="What to do with sampled subgraph (default: delete)",
-)
-@click.option(
-    "--output-file",
-    type=click.Path(dir_okay=False, writable=True),
-    help="Path to export file (for export mode)",
-)
-@click.option(
-    "--output-format",
-    type=click.Choice(["yaml", "json"], case_sensitive=False),
-    default="yaml",
-    help="Export format (default: yaml)",
-)
-@click.option(
-    "--dry-run",
-    is_flag=True,
-    help="Preview changes without executing",
-)
-@click.option(
-    "--no-validate",
-    is_flag=True,
-    help="Skip validation checks",
-)
-@click.option(
-    "--config",
-    type=click.Path(exists=True, dir_okay=False, readable=True),
-    help="Path to scale configuration file",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def scale_down_algorithm(
-    ctx: click.Context,
-    tenant_id: Optional[str],
-    algorithm: str,
-    target_size: float,
-    target_count: Optional[int],
-    burn_in: int,
-    burning_prob: float,
-    walk_length: int,
-    alpha: float,
-    output_mode: str,
-    output_file: Optional[str],
-    output_format: str,
-    dry_run: bool,
-    no_validate: bool,
-    config: Optional[str],
-    no_container: bool,
-) -> None:
-    """
-    Scale down using sampling algorithms.
-
-    Implements graph sampling algorithms that preserve structural properties:
-
-    \b
-    - forest-fire: Simulates forest fire spreading through graph
-    - mhrw: Metropolis-Hastings Random Walk for unbiased sampling
-    - random-walk: Simple random walk sampling
-    - random-node: Uniform random node sampling
-
-    Example:
-
-        \b
-        # Sample 10% of graph using forest-fire
-        atg scale-down algorithm --algorithm forest-fire --target-size 0.1
-
-        \b
-        # Sample exactly 500 nodes using MHRW
-        atg scale-down algorithm --algorithm mhrw --target-count 500
-
-        \b
-        # Export sampled graph to YAML file
-        atg scale-down algorithm --algorithm forest-fire --target-size 0.2 \\
-            --output-mode export --output-file sampled_graph.yaml
-    """
-    from src.cli_commands_scale import scale_down_algorithm_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    if not no_container:
-        ensure_neo4j_running(debug)
-
-    await scale_down_algorithm_command_handler(
-        tenant_id=tenant_id,
-        algorithm=algorithm,
-        target_size=target_size,
-        target_count=target_count,
-        burn_in=burn_in,
-        burning_prob=burning_prob,
-        walk_length=walk_length,
-        alpha=alpha,
-        output_mode=output_mode,
-        output_file=output_file,
-        output_format=output_format,
-        dry_run=dry_run,
-        validate=not no_validate,
-        config_path=config,
-        debug=debug,
-        no_container=no_container,
-    )
-
-
-@scale_down.command(name="pattern")
-@click.option(
-    "--tenant-id",
-    required=False,
-    help="Azure tenant ID (defaults to AZURE_TENANT_ID from .env)",
-)
-@click.option(
-    "--pattern",
-    required=True,
-    type=click.Choice(
-        ["security", "network", "compute", "storage", "resource-type"],
-        case_sensitive=False,
-    ),
-    help="Pattern-based filter to apply",
-)
-@click.option(
-    "--resource-types",
-    type=str,
-    help="Comma-separated resource types (for resource-type pattern)",
-)
-@click.option(
-    "--target-size",
-    type=float,
-    default=0.1,
-    help="Target size as fraction of pattern match (default: 0.1)",
-)
-@click.option(
-    "--output-mode",
-    type=click.Choice(["delete", "export", "new-tenant"], case_sensitive=False),
-    default="delete",
-    help="What to do with sampled subgraph (default: delete)",
-)
-@click.option(
-    "--output-file",
-    type=click.Path(dir_okay=False, writable=True),
-    help="Path to export file (for export mode)",
-)
-@click.option(
-    "--output-format",
-    type=click.Choice(["yaml", "json"], case_sensitive=False),
-    default="yaml",
-    help="Export format (default: yaml)",
-)
-@click.option(
-    "--dry-run",
-    is_flag=True,
-    help="Preview changes without executing",
-)
-@click.option(
-    "--no-validate",
-    is_flag=True,
-    help="Skip validation checks",
-)
-@click.option(
-    "--config",
-    type=click.Path(exists=True, dir_okay=False, readable=True),
-    help="Path to scale configuration file",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def scale_down_pattern(
-    ctx: click.Context,
-    tenant_id: Optional[str],
-    pattern: str,
-    resource_types: Optional[str],
-    target_size: float,
-    output_mode: str,
-    output_file: Optional[str],
-    output_format: str,
-    dry_run: bool,
-    no_validate: bool,
-    config: Optional[str],
-    no_container: bool,
-) -> None:
-    """
-    Scale down using pattern-based filtering.
-
-    Filters graph by specific resource patterns before sampling:
-
-    \b
-    - security: Key Vaults, Network Security Groups, managed identities
-    - network: VNets, subnets, NICs, load balancers, gateways
-    - compute: VMs, VM scale sets, container instances
-    - storage: Storage accounts, disks, file shares
-    - resource-type: Custom list of resource types
-
-    Example:
-
-        \b
-        # Keep only security-related resources
-        atg scale-down pattern --pattern security --target-size 1.0
-
-        \b
-        # Sample network resources (keep 30%)
-        atg scale-down pattern --pattern network --target-size 0.3
-
-        \b
-        # Filter by specific resource types
-        atg scale-down pattern --pattern resource-type \\
-            --resource-types "Microsoft.Compute/virtualMachines,Microsoft.Network/virtualNetworks"
-    """
-    from src.cli_commands_scale import scale_down_pattern_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    if not no_container:
-        ensure_neo4j_running(debug)
-
-    # Parse resource types if provided
-    resource_type_list = resource_types.split(",") if resource_types else None
-
-    await scale_down_pattern_command_handler(
-        tenant_id=tenant_id,
-        pattern=pattern,
-        resource_types=resource_type_list,
-        target_size=target_size,
-        output_mode=output_mode,
-        output_file=output_file,
-        output_format=output_format,
-        dry_run=dry_run,
-        validate=not no_validate,
-        config_path=config,
-        debug=debug,
-        no_container=no_container,
-    )
-
-
-@cli.command(name="scale-clean")
-@click.option(
-    "--tenant-id",
-    required=False,
-    help="Azure tenant ID (defaults to AZURE_TENANT_ID from .env)",
-)
-@click.option(
-    "--force",
-    is_flag=True,
-    help="Skip confirmation prompt",
-)
-@click.option(
-    "--dry-run",
-    is_flag=True,
-    help="Preview what would be deleted without executing",
-)
-@click.option(
-    "--output-format",
-    type=click.Choice(["table", "json", "markdown"], case_sensitive=False),
-    default="table",
-    help="Output format for results (default: table)",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def scale_clean(
-    ctx: click.Context,
-    tenant_id: Optional[str],
-    force: bool,
-    dry_run: bool,
-    output_format: str,
-    no_container: bool,
-) -> None:
-    """
-    Clean up all synthetic data from the graph.
-
-    Removes all nodes with the :Synthetic label and their relationships.
-    This operation is useful for resetting the graph to its original state
-    after scale testing.
-
-    Example:
-
-        \b
-        # Preview what would be cleaned
-        atg scale-clean --dry-run
-
-        \b
-        # Clean synthetic data with confirmation
-        atg scale-clean
-
-        \b
-        # Clean without confirmation
-        atg scale-clean --force
-    """
-    from src.cli_commands_scale import scale_clean_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    if not no_container:
-        ensure_neo4j_running(debug)
-
-    await scale_clean_command_handler(
-        tenant_id=tenant_id,
-        force=force,
-        dry_run=dry_run,
-        output_format=output_format,
-        debug=debug,
-        no_container=no_container,
-    )
-
-
-@cli.command(name="scale-validate")
-@click.option(
-    "--tenant-id",
-    required=False,
-    help="Azure tenant ID (defaults to AZURE_TENANT_ID from .env)",
-)
-@click.option(
-    "--fix",
-    is_flag=True,
-    help="Attempt to fix validation issues automatically",
-)
-@click.option(
-    "--output-format",
-    type=click.Choice(["table", "json", "markdown"], case_sensitive=False),
-    default="table",
-    help="Output format for results (default: table)",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def scale_validate(
-    ctx: click.Context,
-    tenant_id: Optional[str],
-    fix: bool,
-    output_format: str,
-    no_container: bool,
-) -> None:
-    """
-    Validate graph integrity after scale operations.
-
-    Runs comprehensive validation checks:
-
-    \b
-    - Graph structure integrity
-    - Relationship consistency
-    - Synthetic node labeling
-    - Orphaned node detection
-    - ID uniqueness
-    - Dual-graph consistency (Original ↔ Abstracted)
-
-    Example:
-
-        \b
-        # Run validation checks
-        atg scale-validate
-
-        \b
-        # Validate and auto-fix issues
-        atg scale-validate --fix
-
-        \b
-        # Export validation report as JSON
-        atg scale-validate --output-format json > validation_report.json
-    """
-    from src.cli_commands_scale import scale_validate_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    if not no_container:
-        ensure_neo4j_running(debug)
-
-    await scale_validate_command_handler(
-        tenant_id=tenant_id,
-        fix=fix,
-        output_format=output_format,
-        debug=debug,
-        no_container=no_container,
-    )
-
-
-@cli.command(name="scale-stats")
-@click.option(
-    "--tenant-id",
-    required=False,
-    help="Azure tenant ID (defaults to AZURE_TENANT_ID from .env)",
-)
-@click.option(
-    "--detailed",
-    is_flag=True,
-    help="Show detailed statistics including per-type breakdown",
-)
-@click.option(
-    "--output-format",
-    type=click.Choice(["table", "json", "markdown"], case_sensitive=False),
-    default="table",
-    help="Output format for results (default: table)",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def scale_stats(
-    ctx: click.Context,
-    tenant_id: Optional[str],
-    detailed: bool,
-    output_format: str,
-    no_container: bool,
-) -> None:
-    """
-    Show graph statistics and metrics.
-
-    Displays current graph state including:
-
-    \b
-    - Total node and relationship counts
-    - Synthetic vs original resource counts
-    - Node type distribution
-    - Relationship type distribution
-    - Dual-graph integrity metrics
-
-    Example:
-
-        \b
-        # Show basic statistics
-        atg scale-stats
-
-        \b
-        # Show detailed statistics with type breakdown
-        atg scale-stats --detailed
-
-        \b
-        # Export stats as JSON
-        atg scale-stats --output-format json > graph_stats.json
-    """
-    from src.cli_commands_scale import scale_stats_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    if not no_container:
-        ensure_neo4j_running(debug)
-
-    await scale_stats_command_handler(
-        tenant_id=tenant_id,
-        detailed=detailed,
-        output_format=output_format,
-        debug=debug,
-        no_container=no_container,
-    )
+# Register scale command groups and individual commands
+cli.add_command(scale_up_cmd, "scale-up")
+cli.add_command(scale_down_cmd, "scale-down")
+cli.add_command(scale_clean_cmd, "scale-clean")
+cli.add_command(scale_validate_cmd, "scale-validate")
+cli.add_command(scale_stats_cmd, "scale-stats")
 
 
 # ============================================================================
-# Layer Management Commands (Issue #424)
-# ============================================================================
-
-
-@cli.group(name="layer")
-def layer() -> None:
-    """Layer management commands for multi-layer graph projections."""
-    pass
-
-
-@layer.command(name="list")
-@click.option(
-    "--tenant-id",
-    help="Filter by tenant ID",
-)
-@click.option(
-    "--include-inactive/--active-only",
-    default=True,
-    help="Show inactive layers (default: true)",
-)
-@click.option(
-    "--type",
-    "layer_type",
-    type=click.Choice(
-        ["baseline", "scaled", "experimental", "snapshot"], case_sensitive=False
-    ),
-    help="Filter by layer type",
-)
-@click.option(
-    "--sort-by",
-    type=click.Choice(["name", "created_at", "node_count"], case_sensitive=False),
-    default="created_at",
-    help="Sort field (default: created_at)",
-)
-@click.option(
-    "--ascending/--descending",
-    default=False,
-    help="Sort order (default: descending)",
-)
-@click.option(
-    "--format",
-    "format_type",
-    type=click.Choice(["table", "json", "yaml"], case_sensitive=False),
-    default="table",
-    help="Output format (default: table)",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def layer_list(
-    ctx: click.Context,
-    tenant_id: Optional[str],
-    include_inactive: bool,
-    layer_type: Optional[str],
-    sort_by: str,
-    ascending: bool,
-    format_type: str,
-    no_container: bool,
-) -> None:
-    """List all layers with summary information."""
-    from src.cli_commands_layer import layer_list_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    await layer_list_command_handler(
-        tenant_id=tenant_id,
-        include_inactive=include_inactive,
-        layer_type=layer_type,
-        sort_by=sort_by,
-        ascending=ascending,
-        format_type=format_type,
-        no_container=no_container,
-        debug=debug,
-    )
-
-
-@layer.command(name="show")
-@click.argument("layer_id")
-@click.option(
-    "--format",
-    "format_type",
-    type=click.Choice(["text", "json", "yaml"], case_sensitive=False),
-    default="text",
-    help="Output format (default: text)",
-)
-@click.option(
-    "--show-stats",
-    is_flag=True,
-    help="Include detailed statistics",
-)
-@click.option(
-    "--show-lineage",
-    is_flag=True,
-    help="Show parent/child layers",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def layer_show(
-    ctx: click.Context,
-    layer_id: str,
-    format_type: str,
-    show_stats: bool,
-    show_lineage: bool,
-    no_container: bool,
-) -> None:
-    """Show detailed information about a specific layer."""
-    from src.cli_commands_layer import layer_show_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    await layer_show_command_handler(
-        layer_id=layer_id,
-        format_type=format_type,
-        show_stats=show_stats,
-        show_lineage=show_lineage,
-        no_container=no_container,
-        debug=debug,
-    )
-
-
-@layer.command(name="active")
-@click.argument("layer_id", required=False)
-@click.option(
-    "--tenant-id",
-    help="Tenant context (for multi-tenant)",
-)
-@click.option(
-    "--format",
-    "format_type",
-    type=click.Choice(["text", "json"], case_sensitive=False),
-    default="text",
-    help="Output format",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def layer_active(
-    ctx: click.Context,
-    layer_id: Optional[str],
-    tenant_id: Optional[str],
-    format_type: str,
-    no_container: bool,
-) -> None:
-    """Show or set the active layer."""
-    from src.cli_commands_layer import layer_active_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    await layer_active_command_handler(
-        layer_id=layer_id,
-        tenant_id=tenant_id,
-        format_type=format_type,
-        no_container=no_container,
-        debug=debug,
-    )
-
-
-@layer.command(name="create")
-@click.argument("layer_id")
-@click.option(
-    "--name",
-    help="Human-readable name (default: layer_id)",
-)
-@click.option(
-    "--description",
-    help="Layer description",
-)
-@click.option(
-    "--type",
-    "layer_type",
-    type=click.Choice(
-        ["baseline", "scaled", "experimental", "snapshot"], case_sensitive=False
-    ),
-    default="experimental",
-    help="Layer type (default: experimental)",
-)
-@click.option(
-    "--parent-layer",
-    help="Parent layer for lineage",
-)
-@click.option(
-    "--tenant-id",
-    help="Tenant ID",
-)
-@click.option(
-    "--tag",
-    "tags",
-    multiple=True,
-    help="Add tag (multiple allowed)",
-)
-@click.option(
-    "--make-active",
-    is_flag=True,
-    help="Set as active layer",
-)
-@click.option(
-    "--yes",
-    is_flag=True,
-    help="Skip confirmation",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def layer_create(
-    ctx: click.Context,
-    layer_id: str,
-    name: Optional[str],
-    description: Optional[str],
-    layer_type: str,
-    parent_layer: Optional[str],
-    tenant_id: Optional[str],
-    tags: tuple,
-    make_active: bool,
-    yes: bool,
-    no_container: bool,
-) -> None:
-    """Create a new empty layer."""
-    from src.cli_commands_layer import layer_create_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    await layer_create_command_handler(
-        layer_id=layer_id,
-        name=name,
-        description=description,
-        layer_type=layer_type,
-        parent_layer=parent_layer,
-        tenant_id=tenant_id,
-        tags=list(tags),
-        make_active=make_active,
-        yes=yes,
-        no_container=no_container,
-        debug=debug,
-    )
-
-
-@layer.command(name="copy")
-@click.argument("source")
-@click.argument("target")
-@click.option(
-    "--name",
-    help="Name for new layer",
-)
-@click.option(
-    "--description",
-    help="Description for new layer",
-)
-@click.option(
-    "--copy-metadata/--no-copy-metadata",
-    default=True,
-    help="Copy metadata dict from source (default: true)",
-)
-@click.option(
-    "--make-active",
-    is_flag=True,
-    help="Set new layer as active",
-)
-@click.option(
-    "--yes",
-    is_flag=True,
-    help="Skip confirmation",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def layer_copy(
-    ctx: click.Context,
-    source: str,
-    target: str,
-    name: Optional[str],
-    description: Optional[str],
-    copy_metadata: bool,
-    make_active: bool,
-    yes: bool,
-    no_container: bool,
-) -> None:
-    """Copy an entire layer (nodes + relationships)."""
-    from src.cli_commands_layer import layer_copy_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    await layer_copy_command_handler(
-        source=source,
-        target=target,
-        name=name,
-        description=description,
-        copy_metadata=copy_metadata,
-        make_active=make_active,
-        yes=yes,
-        no_container=no_container,
-        debug=debug,
-    )
-
-
-@layer.command(name="delete")
-@click.argument("layer_id")
-@click.option(
-    "--force",
-    is_flag=True,
-    help="Allow deletion of active/baseline layers",
-)
-@click.option(
-    "--yes",
-    is_flag=True,
-    help="Skip confirmation",
-)
-@click.option(
-    "--archive",
-    type=click.Path(dir_okay=False, writable=True),
-    help="Archive layer before deletion",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def layer_delete(
-    ctx: click.Context,
-    layer_id: str,
-    force: bool,
-    yes: bool,
-    archive: Optional[str],
-    no_container: bool,
-) -> None:
-    """Delete a layer and all its nodes/relationships."""
-    from src.cli_commands_layer import layer_delete_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    await layer_delete_command_handler(
-        layer_id=layer_id,
-        force=force,
-        yes=yes,
-        archive=archive,
-        no_container=no_container,
-        debug=debug,
-    )
-
-
-@layer.command(name="diff")
-@click.argument("layer_a")
-@click.argument("layer_b")
-@click.option(
-    "--detailed",
-    is_flag=True,
-    help="Include node IDs in output",
-)
-@click.option(
-    "--properties",
-    is_flag=True,
-    help="Compare property values",
-)
-@click.option(
-    "--output",
-    type=click.Path(dir_okay=False, writable=True),
-    help="Save report to file",
-)
-@click.option(
-    "--format",
-    "format_type",
-    type=click.Choice(["text", "json"], case_sensitive=False),
-    default="text",
-    help="Output format (default: text)",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def layer_diff(
-    ctx: click.Context,
-    layer_a: str,
-    layer_b: str,
-    detailed: bool,
-    properties: bool,
-    output: Optional[str],
-    format_type: str,
-    no_container: bool,
-) -> None:
-    """Compare two layers to find differences."""
-    from src.cli_commands_layer import layer_diff_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    await layer_diff_command_handler(
-        layer_a=layer_a,
-        layer_b=layer_b,
-        detailed=detailed,
-        properties=properties,
-        output=output,
-        format_type=format_type,
-        no_container=no_container,
-        debug=debug,
-    )
-
-
-@layer.command(name="validate")
-@click.argument("layer_id")
-@click.option(
-    "--fix",
-    is_flag=True,
-    help="Attempt automatic fixes",
-)
-@click.option(
-    "--output",
-    type=click.Path(dir_okay=False, writable=True),
-    help="Save report to file",
-)
-@click.option(
-    "--format",
-    "format_type",
-    type=click.Choice(["text", "json"], case_sensitive=False),
-    default="text",
-    help="Output format (default: text)",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def layer_validate(
-    ctx: click.Context,
-    layer_id: str,
-    fix: bool,
-    output: Optional[str],
-    format_type: str,
-    no_container: bool,
-) -> None:
-    """Validate layer integrity and check for issues."""
-    from src.cli_commands_layer import layer_validate_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    await layer_validate_command_handler(
-        layer_id=layer_id,
-        fix=fix,
-        output=output,
-        format_type=format_type,
-        no_container=no_container,
-        debug=debug,
-    )
-
-
-@layer.command(name="refresh-stats")
-@click.argument("layer_id")
-@click.option(
-    "--format",
-    "format_type",
-    type=click.Choice(["text", "json"], case_sensitive=False),
-    default="text",
-    help="Output format",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def layer_refresh_stats(
-    ctx: click.Context,
-    layer_id: str,
-    format_type: str,
-    no_container: bool,
-) -> None:
-    """Refresh layer metadata statistics."""
-    from src.cli_commands_layer import layer_refresh_stats_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    await layer_refresh_stats_command_handler(
-        layer_id=layer_id,
-        format_type=format_type,
-        no_container=no_container,
-        debug=debug,
-    )
-
-
-@layer.command(name="archive")
-@click.argument("layer_id")
-@click.argument("output_path", type=click.Path(dir_okay=False, writable=True))
-@click.option(
-    "--include-original",
-    is_flag=True,
-    help="Include Original nodes in archive",
-)
-@click.option(
-    "--compress",
-    is_flag=True,
-    help="Compress archive with gzip",
-)
-@click.option(
-    "--yes",
-    is_flag=True,
-    help="Skip confirmation",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def layer_archive(
-    ctx: click.Context,
-    layer_id: str,
-    output_path: str,
-    include_original: bool,
-    compress: bool,
-    yes: bool,
-    no_container: bool,
-) -> None:
-    """Export layer to JSON archive file."""
-    from src.cli_commands_layer import layer_archive_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    await layer_archive_command_handler(
-        layer_id=layer_id,
-        output_path=output_path,
-        include_original=include_original,
-        compress=compress,
-        yes=yes,
-        no_container=no_container,
-        debug=debug,
-    )
-
-
-@layer.command(name="restore")
-@click.argument(
-    "archive_path", type=click.Path(exists=True, dir_okay=False, readable=True)
-)
-@click.option(
-    "--layer-id",
-    help="Override layer ID from archive",
-)
-@click.option(
-    "--make-active",
-    is_flag=True,
-    help="Set as active layer after restore",
-)
-@click.option(
-    "--yes",
-    is_flag=True,
-    help="Skip confirmation",
-)
-@click.option(
-    "--no-container",
-    is_flag=True,
-    help="Do not auto-start Neo4j container",
-)
-@click.pass_context
-@async_command
-async def layer_restore(
-    ctx: click.Context,
-    archive_path: str,
-    layer_id: Optional[str],
-    make_active: bool,
-    yes: bool,
-    no_container: bool,
-) -> None:
-    """Restore layer from JSON archive."""
-    from src.cli_commands_layer import layer_restore_command_handler
-
-    debug = ctx.obj.get("debug", False)
-    await layer_restore_command_handler(
-        archive_path=archive_path,
-        layer_id=layer_id,
-        make_active=make_active,
-        yes=yes,
-        no_container=no_container,
-        debug=debug,
-    )
-
-
 def main() -> None:
     """Main entry point."""
     result = cli()  # type: ignore[reportCallIssue]
